Move setup/teardown code into fixtures module
