+++ .gitignore
@@ -1,0 +1,3 @@
+manifests/
+playbooks/
+*.swp
